pyproject.toml
[project]
name = "ordinary-meaning"
version = "0.1.0"
description = "Autoresearch loop for the ordinary meaning of legal terms"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.52.0",
    "openai>=1.82.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "streamlit>=1.56.0",
    "pymupdf>=1.27.2.2",
    "python-docx>=1.2.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
]
